Reduced logging for unknown ICMP types.

+++ include/dp_flow.h
@@ -119,7 +119,7 @@ void dp_add_flow_data(struct flow_key *key, void *data);
 void dp_add_flow(struct flow_key *key);
 void dp_delete_flow_key(struct flow_key *key);
 bool dp_flow_exists(struct flow_key *key);
-int8_t dp_build_flow_key(struct flow_key *key /* out */, struct rte_mbuf *m /* in */);
+int dp_build_flow_key(struct flow_key *key /* out */, struct rte_mbuf *m /* in */);
 void dp_invert_flow_key(struct flow_key *key /* in / out */);
 int dp_flow_init(int socket_id);
 void dp_flow_free();
